Organizing routes into a fastapi router
